Commiting some things so I can update. (I need the token equals() to test)
